Move BoardArtist to its own file board_artist.py
